refactor: MVVM based file structure (#34)
